Issue #5104: created internal utils package for test utils
