+++ config/suppressions.xml
@@ -26,7 +26,7 @@
     <!-- Tone down the checking for test code -->
     <suppress checks="CyclomaticComplexity" files="[\\/]XdocsPagesTest\.java"/>
     <suppress checks="NPathComplexity" files="[\\/]XdocsPagesTest\.java"/>
-    <suppress checks="IllegalCatch" files="[\\/]internal[\\/]\w+Util\.java"/>
+    <suppress checks="IllegalCatch" files="[\\/]internal[\\/].*[\\/]\w+Util\.java"/>
     <suppress checks="EmptyBlock" files=".*[\\/]src[\\/]test[\\/]"/>
     <suppress checks="ImportControl" files=".*[\\/]src[\\/](test|it)[\\/]"/>
     <suppress checks="Javadoc" files=".*[\\/]src[\\/](test|it)[\\/]"/>
