+++ src/test/java/com/puppycrawl/tools/checkstyle/BaseCheckTestSupport.java
@@ -41,6 +41,7 @@ import com.google.common.collect.MapDifference;
 import com.google.common.collect.MapDifference.ValueDifference;
 import com.google.common.collect.Maps;
 import com.puppycrawl.tools.checkstyle.api.Configuration;
+import com.puppycrawl.tools.checkstyle.internal.utils.BriefUtLogger;
 
 /**
  * @deprecated BaseCheckTestSupport is no longer used. All checks
