+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/sizes/ExecutableStatementCountCheckTest.java
@@ -34,7 +34,7 @@ import com.puppycrawl.tools.checkstyle.DefaultConfiguration;
 import com.puppycrawl.tools.checkstyle.api.Context;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
-import com.puppycrawl.tools.checkstyle.internal.TestUtils;
+import com.puppycrawl.tools.checkstyle.internal.utils.TestUtil;
 import com.puppycrawl.tools.checkstyle.utils.CommonUtils;
 
 public class ExecutableStatementCountCheckTest
@@ -51,7 +51,7 @@ public class ExecutableStatementCountCheckTest
         ast.setType(TokenTypes.STATIC_INIT);
         final ExecutableStatementCountCheck check = new ExecutableStatementCountCheck();
         Assert.assertTrue("Stateful field is not cleared after beginTree",
-                TestUtils.isStatefulFieldClearedDuringBeginTree(check, ast, "contextStack",
+                TestUtil.isStatefulFieldClearedDuringBeginTree(check, ast, "contextStack",
                     contextStack -> ((Collection<Context>) contextStack).isEmpty()));
     }
