+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/utils/XdocUtil.java
@@ -17,7 +17,7 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ////////////////////////////////////////////////////////////////////////////////
 
-package com.puppycrawl.tools.checkstyle.internal;
+package com.puppycrawl.tools.checkstyle.internal.utils;
 
 import java.io.IOException;
 import java.nio.file.DirectoryStream;
@@ -37,6 +37,7 @@ import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
+/** @noinspection ClassOnlyUsedInOnePackage */
 public final class XdocUtil {
     public static final String DIRECTORY_PATH = "src/xdocs";
