+++ src/test/java/com/puppycrawl/tools/checkstyle/xpath/ElementNodeTest.java
@@ -19,7 +19,7 @@
 
 package com.puppycrawl.tools.checkstyle.xpath;
 
-import static com.puppycrawl.tools.checkstyle.internal.XpathUtil.getXpathItems;
+import static com.puppycrawl.tools.checkstyle.internal.utils.XpathUtil.getXpathItems;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -32,7 +32,7 @@ import org.junit.Test;
 import com.puppycrawl.tools.checkstyle.AbstractPathTestSupport;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
-import com.puppycrawl.tools.checkstyle.internal.TestUtils;
+import com.puppycrawl.tools.checkstyle.internal.utils.TestUtil;
 import net.sf.saxon.om.Item;
 import net.sf.saxon.om.NodeInfo;
 
@@ -48,7 +48,7 @@ public class ElementNodeTest extends AbstractPathTestSupport {
     @Before
     public void init() throws Exception {
         final File file = new File(getPath("InputXpathMapperAst.java"));
-        final DetailAST rootAst = TestUtils.parseFile(file);
+        final DetailAST rootAst = TestUtil.parseFile(file);
         rootNode = new RootNode(rootAst);
     }
