+++ src/test/java/com/puppycrawl/tools/checkstyle/xpath/XpathMapperTest.java
@@ -19,7 +19,7 @@
 
 package com.puppycrawl.tools.checkstyle.xpath;
 
-import static com.puppycrawl.tools.checkstyle.internal.XpathUtil.getXpathItems;
+import static com.puppycrawl.tools.checkstyle.internal.utils.XpathUtil.getXpathItems;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -34,7 +34,7 @@ import org.junit.Test;
 import com.puppycrawl.tools.checkstyle.AbstractPathTestSupport;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
-import com.puppycrawl.tools.checkstyle.internal.TestUtils;
+import com.puppycrawl.tools.checkstyle.internal.utils.TestUtil;
 import net.sf.saxon.om.AxisInfo;
 import net.sf.saxon.om.Item;
 import net.sf.saxon.trans.XPathException;
@@ -529,7 +529,7 @@ public class XpathMapperTest extends AbstractPathTestSupport {
 
     private RootNode getRootNode(String fileName) throws Exception {
         final File file = new File(getPath(fileName));
-        final DetailAST rootAst = TestUtils.parseFile(file);
+        final DetailAST rootAst = TestUtil.parseFile(file);
         return new RootNode(rootAst);
     }
